iter5 correction withdraw get

+++ internal/handlers/withdraw/request.go
@@ -2,5 +2,5 @@ package withdraw
 
 type RequestBody struct {
 	Order string  `json:"order"`
-	Sum   float64 `json:"sum"`
+	Sum   float32 `json:"sum"`
 }
